Law_chatbot_project_workshop3/requirements.txt
flask
openai
python-dotenv
faiss-cpu
PyMuPDF